Removed connect.js, fixed all bugs so far in the cli because of that
